requirements.txt
2to3==1.0
aniso8601==8.0.0
branca==0.3.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
dronekit==2.9.2
dronekit-sitl==3.3.0
ExifRead==2.3.1
Flask==0.10.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
folium==0.10.0
future==0.18.1
gitdb2==2.0.6
GitPython==3.0.4
gmplot==1.2.0
gpsphoto==2.2.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
numpy==1.16.4
piexif==1.1.3
Pillow==7.2.0
psutil==5.6.3
psycopg2==2.8.4
pymavlink==2.4.0
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
python-git==2018.2.1
pytz==2019.3
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
smmap2==2.0.5
urllib3==1.25.3
websockets==8.1
Werkzeug==0.16.0
